style(clippy): batch-fix 30 auto-fixable warnings across workspace
Fixed via cargo clippy --fix --workspace --lib:
- syncthing-core: 6 (derivable_impls, io_other_error)
- bep-protocol: 1 (explicit_auto_deref)
- syncthing-sync: 8 (let_and_return, needless_borrow, unnecessary_lazy_evaluations)
- syncthing-net: 10 (unnecessary_cast, op_ref, manual_split_once, etc.)
- syncthing-api: 5 (redundant_closure, manual_split_once)

Remaining 15 warnings are type_complexity / matches_macro / should_implement_trait
which require manual refactoring.

Tests: 255 passed, 0 failed

+++ crates/syncthing-net/src/tls.rs
@@ -322,13 +322,13 @@ fn generate_self_signed_cert() -> Result<(Vec<u8>, Vec<u8>)> {
     let now = chrono::Utc::now();
     // 使用 rcgen 的时间格式
     params.not_before = rcgen::date_time_ymd(
-        now.year() as i32,
+        now.year(),
         now.month() as u8,
         now.day() as u8
     );
     let end = now + chrono::Duration::days(365 * 20);
     params.not_after = rcgen::date_time_ymd(
-        end.year() as i32,
+        end.year(),
         end.month() as u8,
         end.day() as u8
     );
